refactor: remove events
