+++ tests/game/test_end_bidding.py
@@ -47,6 +47,7 @@ class TestEndBidding(TestCase):
 
         # old round ended as completed no bidders
         self.assertEqual(game.rounds[-2].status, RoundStatus.COMPLETED_NO_BIDDERS)
+        self.assertTrue(game.rounds[-2].completed)
         # new round in bidding
         self.assertEqual(game.status, RoundStatus.BIDDING)
